+++ backend/friend_request/tests/friend_request_handler_test.go
@@ -1,0 +1,1 @@
+package friend_request
