+++ backend/friend_request/tests/friend_request_operations_test.go
@@ -1,0 +1,1 @@
+package friend_request
