+++ backend/friends/friends_operations.go
@@ -1,19 +1,12 @@
-package handle_friend
+package friends
 
 import (
 	"fmt"
 	"log"
-	"net/http"
 	"strings"
-	"take-a-break/web-service/auth"
 	"take-a-break/web-service/database"
 	"take-a-break/web-service/models"
-
-	"github.com/gin-gonic/gin"
 )
-
-type User = models.User
-type UserRequest = models.UserRequest
 
 // SearchFriends searches for friends based on username and/or name.
 // It returns a slice of User structs matching the search criteria.
@@ -23,6 +16,7 @@ func SearchFriends(conn *database.DBConnection, searchTerm string, emailID strin
 	SELECT 
         u.email_id,
         u.name,
+		u.avatar,
         CASE 
           WHEN friend.email_id1 IS NOT NULL THEN 1 
           WHEN fr.sender IS NOT NULL THEN 2
@@ -84,43 +78,37 @@ func DeleteFriend(conn *database.DBConnection, emailID1 string, emailID2 string)
 	return nil
 }
 
-func SearchFriendsHandler(conn *database.DBConnection) gin.HandlerFunc {
-	return func(c *gin.Context) {
+func FetchFriends(conn *database.DBConnection, email_id string) ([]models.User, error) {
+	query := `
+		SELECT u.email_id, u.name, u.role, u.avatar
+		FROM users u
+		INNER JOIN friends f
+		ON u.email_id = f.email_id2
+		WHERE f.email_id1 = $1
+		ORDER BY u.name
+	`
 
-		token := c.Param("token")
-		claims := auth.ReturnJWTToken(token)
-		emailID := claims["email"].(string)
-
-		searchTerm := c.Query("searchTerm")
-		foundUsers, err := SearchFriends(conn, searchTerm, emailID)
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Error searching for friends"})
-			return
-		}
-		var friendCards []gin.H
-		for _, user := range foundUsers {
-			friendCard := gin.H{
-				"name":             user.Name,
-				"email":            user.EmailID,
-				"has_sent_request": user.SentRequest,
-				"image":            "./UCSD-logo.png",
-			}
-			friendCards = append(friendCards, friendCard)
-		}
-
-		c.JSON(http.StatusOK, friendCards)
+	rows, err := conn.ExecuteQuery(query, email_id)
+	if err != nil {
+		return nil, err
 	}
-}
+	defer rows.Close()
 
-func DeleteFriendHandler(conn *database.DBConnection) gin.HandlerFunc {
-	return func(c *gin.Context) {
-		emailID1 := c.PostForm("email_id1")
-		emailID2 := c.PostForm("email_id2")
-		err := DeleteFriend(conn, emailID1, emailID2)
+	var friends []models.User
+	for rows.Next() {
+		var friend models.User
+		err := rows.Scan(
+			&friend.EmailID,
+			&friend.Name,
+			&friend.Role,
+			&friend.Avatar,
+		)
 		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Error deleting friend"})
-			return
+			return nil, err
 		}
-		c.JSON(http.StatusOK, gin.H{"message": fmt.Sprintf("Friendship between '%s' and '%s' deleted successfully", emailID1, emailID2)})
+
+		friends = append(friends, friend)
 	}
+
+	return friends, nil
 }
