+++ backend/friends/tests/friends_operations_test.go
@@ -1,0 +1,1 @@
+package friends
