+++ backend/models/models.go
@@ -29,6 +29,7 @@ type UserRequest struct {
 	EmailID     string `json:"email_id"`
 	Name        string `json:"name"`
 	SentRequest string `json:"has_sent_request"`
+	Avatar      string `json:"avatar"`
 }
 
 type Config struct {
