+++ backend/users/users_operations.go
@@ -7,9 +7,9 @@ import (
 
 func InsertUserIntoDatabase(conn *database.DBConnection, user User) (User, error) {
 	insertQuery := `
-		INSERT INTO users (email_id, name, role)
+		INSERT INTO users (email_id, name, role, avatar)
 		VALUES ($1, $2, $3) RETURNING
-		email_id, name, role
+		email_id, name, role, avatar
 	`
 
 	// Execute the INSERT query using the ExecuteQuery function
@@ -18,6 +18,7 @@ func InsertUserIntoDatabase(conn *database.DBConnection, user User) (User, error
 		user.EmailID,
 		user.Name,
 		user.Role,
+		user.Avatar,
 	)
 
 	if err != nil {
@@ -30,6 +31,7 @@ func InsertUserIntoDatabase(conn *database.DBConnection, user User) (User, error
 			&newUser.EmailID,
 			&newUser.Name,
 			&newUser.Role,
+			&newUser.Avatar,
 		)
 
 		if err != nil {
@@ -59,6 +61,7 @@ func FetchUserByEmailID(conn *database.DBConnection, email_id string) (User, err
 			&newUser.EmailID,
 			&newUser.Name,
 			&newUser.Role,
+			&newUser.Avatar,
 		)
 		if err != nil {
 			return User{}, err
@@ -68,131 +71,4 @@ func FetchUserByEmailID(conn *database.DBConnection, email_id string) (User, err
 	}
 
 	return User{}, errors.New("user not found")
-}
-
-func MakeFriends(conn *database.DBConnection, user1_email, user2_email string) error {
-	query := `
-		INSERT INTO friends (email_id1, email_id2)
-		VALUES ($1, $2), ($3, $4)
-	`
-	// make a bidirectional connection
-	_, err := conn.ExecuteQuery(query, user1_email, user2_email, user2_email, user1_email)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func SendFriendRequest(conn *database.DBConnection, user1_email, user2_email string) error {
-	query := `
-		INSERT INTO friend_requests (sender, reciever)
-		VALUES ($1, $2)
-	`
-	_, err := conn.ExecuteQuery(query, user1_email, user2_email)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func AcceptFriendRequest(conn *database.DBConnection, user1_email, user2_email string) error {
-	// Delete the friend request from the table
-	query := `
-		DELETE FROM friend_requests
-		WHERE (sender = $1 AND reciever = $2)
-	`
-	_, err := conn.ExecuteQuery(query, user1_email, user2_email)
-	if err != nil {
-		return err
-	}
-
-	err = MakeFriends(conn, user1_email, user2_email)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func IgnoreFriendRequest(conn *database.DBConnection, user1_email, user2_email string) error {
-	query := `
-		UPDATE friend_requests
-		SET ignored = true
-		WHERE (sender = $1 AND reciever = $2)
-	`
-	_, err := conn.ExecuteQuery(query, user1_email, user2_email)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func FetchFriends(conn *database.DBConnection, email_id string) ([]User, error) {
-	query := `
-		SELECT u.email_id, u.name, u.role
-		FROM users u
-		INNER JOIN friends f
-		ON u.email_id = f.email_id2
-		WHERE f.email_id1 = $1
-		ORDER BY u.name
-	`
-
-	rows, err := conn.ExecuteQuery(query, email_id)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var friends []User
-	for rows.Next() {
-		var friend User
-		err := rows.Scan(
-			&friend.EmailID,
-			&friend.Name,
-			&friend.Role,
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		friends = append(friends, friend)
-	}
-
-	return friends, nil
-}
-
-func FetchFriendRequest(conn *database.DBConnection, email_id string) ([]User, error) {
-	query := `
-		SELECT u.email_id, u.name, u.role
-		FROM users u
-		INNER JOIN friend_requests fr
-		ON u.email_id = fr.sender
-		WHERE fr.reciever = $1 AND fr.ignored = false
-	`
-
-	rows, err := conn.ExecuteQuery(query, email_id)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var friends []User
-	for rows.Next() {
-		var friend User
-		err := rows.Scan(
-			&friend.EmailID,
-			&friend.Name,
-			&friend.Role,
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		friends = append(friends, friend)
-	}
-
-	return friends, nil
 }
